+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # ChangeLog
 
-* fix fluid filter require specific amount
+* fix crash that some stupid mods try to drain the `EMPTY` fluid